requirements.txt
fastapi
uvicorn
pyyaml
tortoise-orm
aiosqlite
